Removes menu id observable

+++ src/app/core/services/menu.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject, combineLatest, Observable, of } from 'rxjs';
+import { combineLatest, Observable, of } from 'rxjs';
 import { concatMap, first, map, shareReplay, switchMap, tap } from 'rxjs/operators';
 
 import { Endpoint } from '@models/enums/endpoint.enum';
@@ -11,9 +11,9 @@ import { lower } from '@shared/utility/format';
 import { sort } from '@shared/utility/sort';
 import { getDays } from '@utility/get-days';
 import { FirestoreService } from './firestore.service';
-import { LocalStorageService } from './local-storage.service';
 import { BatchService } from './batch.service';
 import { DishService } from './dish.service';
+import { LocalStorageService } from './local-storage.service';
 import { UserService } from './user.service';
 
 @Injectable({
@@ -21,7 +21,6 @@ import { UserService } from './user.service';
 })
 export class MenuService {
   private _endpoint = Endpoint.menus;
-  private _menuId$ = new BehaviorSubject<string>('');
 
   constructor(
     private _firestoreService: FirestoreService,
@@ -31,46 +30,13 @@ export class MenuService {
     private _userService: UserService,
   ) { }
 
-  public get menuId$(): Observable<string> {
-    return this._menuId$.pipe(
-      shareReplay({ refCount: true, bufferSize: 1 })
-    );
-  }
-
-  public updateSavedMenuId(routeMenuId?: string): Observable<string> {
-    const savedMenuId = this._localStorageService.getMenuId();
-    if (routeMenuId) {
-      return of(routeMenuId).pipe(
-        first(),
-        tap(routeMenuId => this._setMenuId(routeMenuId)),
-      );
-    } else if (savedMenuId) {
-      return of(savedMenuId).pipe(
-        first(),
-        tap(savedMenuId => this._setMenuId(savedMenuId))
-      );
-    } else {
-      return this.getMenus().pipe(
-        first(),
-        map(menus => menus[0]?.id ?? ''),
-        tap(firstMenuId => this._setMenuId(firstMenuId))
-      );
-    }
-  }
-
-  public getMenu(): Observable<Menu | undefined> {
-    return this._menuId$.pipe(
-      switchMap(id => {
-        if (!id) {
+  public getMenu(id: string): Observable<Menu | undefined> {
+    return this._firestoreService.getOne<MenuDto>(this._endpoint, id).pipe(
+      switchMap(menuDto => {
+        if (!menuDto) {
           return of(undefined);
         }
-        return this._firestoreService.getOne<MenuDto>(this._endpoint, id);
-      }),
-      switchMap(menuDbo => {
-        if (!menuDbo) {
-          return of(undefined);
-        }
-        return this._transformMenuDbo(menuDbo);
+        return this._transformMenuDto(menuDto);
       })
     );
   }
@@ -79,7 +45,7 @@ export class MenuService {
     return this._userService.uid$.pipe(
       switchMap(uid => this._firestoreService.getMany<MenuDto>(this._endpoint, uid)),
       switchMap(menus => combineLatest(
-        menus.map(menu => this._transformMenuDbo(menu))
+        menus.map(menu => this._transformMenuDto(menu))
       )),
       map(menus => sort(menus, menu => lower(menu.name))),
     );
@@ -108,50 +74,34 @@ export class MenuService {
     return this._updateMenu(id, { name });
   }
 
-  public updateMenuContents({ day, dishId, selected }: {
+  public updateMenuContents({ menu, day, dishId, selected }: {
+    menu: Menu,
     day: Day,
     dishId: string,
     selected: boolean,
-  }): Observable<Menu | undefined> {
-    return this.getMenu().pipe(
-      first(),
-      tap(async menu => {
-        if (!menu) {
-          return;
-        }
-        this._batchService.updateMenuContents({ menu, day, dishId, selected });
-      })
-    );
+  }): Promise<void> {
+    return this._batchService.updateMenuContents({ menu, day, dishId, selected });
   }
 
   public async deleteMenu(id?: string): Promise<void> {
     if (id) {
-      this._firestoreService.getOne<MenuDto>(this._endpoint, id).pipe(
+      this.getMenu(id).pipe(
         first(),
         tap(async menu => {
           if (!menu) {
             return;
           }
           await this._batchService.deleteMenu(menu);
+          if (id === this._localStorageService.getMenuId()) {
+            this._localStorageService.deleteMenuId();
+          }
         })
       ).subscribe();
     }
-    this._localStorageService.deleteMenuId();
-    this.updateSavedMenuId().pipe(
-      first()
-    ).subscribe();
   }
 
-  public deleteMenuContents(day?: Day) {
-    return this.getMenu().pipe(
-      first(),
-      tap(async menu => {
-        if (!menu) {
-          return;
-        }
-        await this._batchService.deleteMenuContents(menu, day);
-      }),
-    );
+  public deleteMenuContents(menu: Menu, day?: Day): Promise<void> {
+    return this._batchService.deleteMenuContents(menu, day);
   }
 
   public getOrderedDays(): Observable<Day[]> {
@@ -161,7 +111,7 @@ export class MenuService {
     );
   }
 
-  private _transformMenuDbo(menu: MenuDto): Observable<Menu> {
+  private _transformMenuDto(menu: MenuDto): Observable<Menu> {
     return combineLatest([
       this.getOrderedDays(),
       this._dishService.getDishes(),
@@ -183,10 +133,5 @@ export class MenuService {
 
   private async _updateMenu(id: string, updates: Partial<MenuDto>): Promise<void> {
     return await this._firestoreService.update<MenuDto>(this._endpoint, id, updates);
-  }
-
-  private _setMenuId(id: string): void {
-    this._localStorageService.setMenuId(id);
-    this._menuId$.next(id);
   }
 }
